Add container build workflow (#1)
Create two Dockerfiles; one of them is for a container with Renode pre-installed, while the other contains the build environment to compile Renode inside of.  
Containers are building Renode with .NET on Ubuntu 24.04.
Also add the Github Actions workflow to build these containers; it is set to upload renode to siliconlabsinc/renode_container only when run from the main branch.

Co-authored-by: yulina.shu <yulina.shu@silabs.com>